refactor(rust): carve v1 wire-format seams to prep for v2
Preparation for the MLT v2 wire format (#1201, docs/migrating-to-v2.md),
with no wire-format or behavior change:

- Convert RawPresence from a tuple struct over Option<RawStream> into an
  enum (AllPresent / Stream) and centralize all presence decoding behind
  RawPresence::decode_bits / decode_bools. v2 can then add raw-bitfield
  and shared-presence-group variants without touching downstream code.
- Move the v1 stream-header wire codec (stream_type byte, encoding byte,
  counts, RLE/Morton extras) into decoder/stream/header01.rs, absorbing
  the stream_type byte codec from physical.rs. StreamMeta / RawStream /
  IntEncoding constructors move to model.rs as format-independent
  in-memory types, so a future header02 codec can sit alongside.
- Rename parse_struct_children to parse_shared_dict_children to match
  shared-dictionary terminology used elsewhere (per #1201).
