+++ rust/mlt-core/src/decoder/stream/mod.rs
@@ -1,6 +1,5 @@
 mod analyze;
 mod decode;
+mod header01;
 pub(crate) mod logical;
 pub(crate) mod model;
-mod parse;
-mod physical;
